requirements.txt
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.4
Flask==2.2.2
idna==3.4
ifaddr==0.1.7
importlib-metadata==5.0.0
imutils==0.5.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
more-itertools==9.0.0
mypy==0.991
mypy-extensions==0.4.3
netdisco==3.0.0
numpy==1.23.4
opencv-python==4.5.4.60
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
py==1.11.0
pycoral==2.0.0
pyparsing==3.0.9
pyserial==3.5
pytest==7.2.0
python-dateutil==2.8.2
requests==2.28.1
scipy==1.9.3
six==1.16.0
tflite-runtime==2.5.0.post1
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.4.0
upnpclient==1.0.3
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
zeroconf==0.39.4
zipp==3.10.0
